The Pinisher game - Put on Backburner
2 Fun little modular scripts.
Started 2 game ideas because I couldnt figure out which one to do. The Pinisher game will be more complicated, so ill put it on hold for now and focus on a very small scope for the space game, and do this game if i have time to do both, or do it next year.